♻️ Refactor: Change the file structure related to the model
